fix policy

+++ app/category/controller.js
@@ -3,6 +3,13 @@ const Category = require('./model');
 
 async function store(req, res, next){
     try {
+        let policy = policyFor(req.user);
+        if(!policy.can('create','Category')){
+         return res.json({
+             error:1,
+             message:`Anda tidak memiliki akses untuk membuat Category`
+         });
+        }
         let payload = req.body;
 
         let category = new Category(payload);
@@ -23,6 +30,13 @@ async function store(req, res, next){
 
 async function update(req, res, next){
     try {
+        let policy = policyFor(req.user);
+        if(!policy.can('update','Category')){
+         return res.json({
+             error:1,
+             message:`Anda tidak memiliki akses untuk mengupdate Category`
+         });
+        }
         let payload = req.body;
 
         let category = await Category.findOneAndUpdate({_id:req.params.id},payload,{new:true,runValidators:true});
@@ -41,7 +55,15 @@ async function update(req, res, next){
 }
 
 async function destroy(req, res, next){
+ 
     try {
+        let policy = policyFor(req.user);
+        if(!policy.can('delete','Category')){
+         return res.json({
+             error:1,
+             message:`Anda tidak memiliki akses untuk membuat Category`
+         });
+        }
         //cari dan hapus category di mongoDB berdasarkan field _id
         let deleted = await Category.findOneAndDelete({
             _id:req.params.id
